kleaf: ddk_module_config.override_parent
This is an attribute that is shipped with the ddk_module_config.parent
feature.

By default, the value is set to "deny", meaning that it is not allowed
to have a ddk_module adding additional kconfig/defconfig's that impacts
the final .config, if it uses a ddk_config. This ensures the best
performance.

For testing purposes, the value may be set to "expect_override",
expecting the ddk_module to add extra defconfig/kconfig's. Note that
the performance may be impaired if they do this, and the module might
be better off NOT using ddk_config in the first place.

We could also add an `allow` value in the future with the least
restrictions if there is a request to add it. Because it is unusual to
set the value to `allow`, we don't allow `allow` in this patch now.

Bug: 370681291
Change-Id: Ic65f6bfc77d72db2682cdab55a2bbf0283da3e51