kleaf: Support --experimental_optimize_ddk_config_actions.
In summary, I invented this --experimental_optimize_ddk_config_actions
flag. With this flag:

* For ddk_module_config -> kernel_build, if the ddk_module_config does
  not have extra defconfig/kconfigs, I deleted the extra DdkConfig
  action, so _kernel_module() just gets .config from kernel_build
  directly.

* For ddk_module_config -> ddk_config -> kernel_build:

  * If the ddk_module_config does not have extra defconfig/kconfigs,
    I deleted the extra DdkConfig action, so _kernel_module() restores
    parent's .config directly.
  * If the ddk_module_config does have extra defconfig/kconfigs, a build
    error is raised, similar to the original behavior.

This is based on the fact that the majority of these DdkConfig actions
does nothing, but still requires creating a sandbox of several tens
of thousands of files, which takes a long time. Instead, we pay a little
price in the analysis phase to do the depset comparison to see if there
are extra defconfig/kconfig, and skips creating this action in the first
place.

The depset comparison for kconfig/defconfig files, though it may incur
depset expansions, should still be very fast. This is because:

* The number of ddk_module() shouldn't exceed a few hundreds for a given
  build
* For each ddk_module, the number of kconfig / defconfig files shouldn't
  exceed a handful on average.

So the number of elements being iterated in depset expansions shouldn't
exceed 10K, which takes a split of a second in the analysis phase on
my machine.

However, we saved N sandboxes, where N = number of ddk_module() without
any extra kconfig / defconfig on top of parent config or kernel_build.
This earns us 30 seconds on my machine on a certain device kernel build.

Bug: 400799412
Change-Id: I37d01e29f24d4d276693db0f933b91447d6f3a50